seperate adc2fno class from radc/uadc to an individual file
