+++ pyscf/adc/uadc.py
@@ -940,210 +940,6 @@ class UADC(lib.StreamObject):
 
         return (list_rdm1_a, list_rdm1_b)
 
-
-class UFNOADC(UADC):
-    #J. Chem. Phys. 159, 084113 (2023)
-    _keys = UADC._keys | {'delta_e','delta_e_corr','e_can','v_can','e_corr_can',
-                          'mo_energy','rdm1_ss','ref_state','trans_guess'
-                          }
-
-    def __init__(self, mf, frozen=0, mo_coeff=None, mo_occ=None):
-        super().__init__(mf, frozen, mo_coeff, mo_occ)
-        self.delta_e = None
-        self.delta_e_corr = None
-        self.mo_energy = None
-        self.method = "adc(3)"
-        self.e_can = None
-        self.v_can = None
-        self.e_corr_can = None
-        self.rdm1_ss = None
-        self.ref_state = None
-        self.if_naf = True
-        self.trans_guess = False
-
-
-    def compute_correction(self, mf, frozen, nroots, eris=None, guess=None):
-        if getattr(self, 'with_df', None) or getattr(self._scf, 'with_df', None):
-            if self.with_df is None:
-                self.with_df = self._scf.with_df
-        adc2_ssfno = UADC(mf, frozen, self.mo_coeff, mo_energy = self.mo_energy, f_ov = self.f_ov).set(
-                                                        verbose = self.verbose,method_type = self.method_type,
-                                                        with_df = self.with_df,if_naf = self.if_naf,
-                                                        thresh_naf = self.thresh_naf,naux = self.naux,
-                                                        ncvs = self.ncvs,
-                                                        approx_trans_moments = self.approx_trans_moments,
-                                                        conv_tol = self.conv_tol,tol_residual = self.tol_residual,
-                                                        max_space = self.max_space, max_cycle = self.max_cycle)
-        e2_ssfno,_,_,_ = adc2_ssfno.kernel(nroots, eris = eris, guess = guess)
-        self.delta_e = self.e_can - e2_ssfno
-        self.delta_e_corr = self.e_corr_can - adc2_ssfno.e_corr
-
-    def kernel(self, nroots=1, guess=None, eris=None, thresh = 1e-4, ref_state = None):
-        cput0 = (logger.process_clock(), logger.perf_counter())
-        log = logger.Logger(self.stdout, self.verbose)
-        self.ref_state = ref_state
-        self.naux = None
-        self.if_heri_eris = True
-        if ref_state is None:
-            logger.info(self,"Do fno adc calculation")
-        elif isinstance(ref_state, int) and 0<ref_state<=nroots:
-            logger.info(self,f"Do ss-fno adc calculation, the specic state is {ref_state}")
-        else:
-            raise ValueError("ref_state should be an int type and in (0,nroots]")
-
-        if not getattr(self, 'with_df', None) and not getattr(self._scf, 'with_df', None):
-                self.if_naf = False
-
-        self.make_ss_rdm1(nroots, self.ref_state, guess)
-        log.timer('make ss rdm1', *cput0)
-        if isinstance(self._scf, scf.rohf.ROHF):
-            self.mo_coeff,self.mo_energy,frozen,self.f_ov = self.make_fno(self.rdm1_ss, thresh)
-        else:
-            self.mo_coeff,self.mo_energy,frozen = self.make_fno(self.rdm1_ss, thresh)
-        log.timer('get frozen info', *cput0)
-        adc3_ssfno = UADC(self._scf, frozen, self.mo_coeff, mo_energy = self.mo_energy, f_ov = self.f_ov).set(
-                                                                    verbose = self.verbose,
-                                                                    method_type = self.method_type,method = self.method,
-                                                                    with_df = self.with_df,
-                                                                    if_naf = self.if_naf,thresh_naf = self.thresh_naf,
-                                                                    if_heri_eris = self.if_heri_eris,ncvs = self.ncvs,
-                                                                    approx_trans_moments = self.approx_trans_moments,
-                                                                    conv_tol = self.conv_tol,
-                                                                    tol_residual = self.tol_residual,
-                                                                    max_space = self.max_space, max_cycle = self.max_cycle)
-
-        if self.if_naf:
-            if self.trans_guess:
-                e_exc, v_exc, spec_fac, x, eris, self.naux = adc3_ssfno.kernel(nroots, guess=self.v_can, eris=eris)
-            else:
-                e_exc, v_exc, spec_fac, x, eris, self.naux = adc3_ssfno.kernel(nroots, guess, eris)
-        else:
-            if self.trans_guess:
-                e_exc, v_exc, spec_fac, x, eris = adc3_ssfno.kernel(nroots, guess=self.v_can, eris=eris)
-            else:
-                e_exc, v_exc, spec_fac, x, eris = adc3_ssfno.kernel(nroots, guess, eris)
-
-        self.e_corr = adc3_ssfno.e_corr
-        log.timer(f'ADC{self.method[4]}', *cput0)
-
-        self.compute_correction(self._scf, frozen, nroots, eris, guess=v_exc)
-        e_exc = e_exc + self.delta_e
-        self.e_corr = self.e_corr + self.delta_e_corr
-
-        msg = ("\n*************************************************************"
-                "\n                   ADC calculation summary"
-                "\n*************************************************************")
-        logger.info(self, msg)
-        logger.info(self, 'FNO MP%s correlation energy of reference state (a.u.) = %.8f',
-                    self.method[4], self.e_corr)
-
-        for n in range(nroots):
-            print_string = ('%s root %d  |  Energy (Eh) = %14.10f  |  Energy (eV) = %12.8f  ' %
-                            (self.method, n, e_exc[n], e_exc[n]*HARTREE2EV))
-            if self.compute_properties:
-                if (self.method_type == "ee"):
-                    print_string += ("|  Osc. strength = %10.8f  " % spec_fac[n])
-                else:
-                    print_string += ("|  Spec. factor = %10.8f  " % spec_fac[n])
-            logger.info(self, print_string)
-            log.timer('UFNOADC', *cput0)
-        return e_exc, v_exc, spec_fac, x
-
-    def make_ss_rdm1(self,nroots,ref_state,guess):
-        adc2_can = UADC(self._scf,self.frozen).set(verbose = self.verbose,method_type = self.method_type,
-                        with_df = self.with_df,if_naf = self.if_naf,thresh_naf = self.thresh_naf,
-                        ncvs = self.ncvs, approx_trans_moments = self.approx_trans_moments,
-                        conv_tol = self.conv_tol,tol_residual = self.tol_residual,
-                        max_space = self.max_space, max_cycle = self.max_cycle)
-        self.e_can,self.v_can,_,_ = adc2_can.kernel(nroots,guess)
-        rdm1_gs = adc2_can.make_ref_rdm1()
-        self.e_corr_can = adc2_can.e_corr
-        if ref_state is not None:
-            rdm1_gs_a = rdm1_gs[0]
-            rdm1_gs_b = rdm1_gs[1]
-            rdm1_ref = adc2_can.make_ref_rdm1()
-            rdm1_ref_a = rdm1_ref[0][ref_state - 1]
-            rdm1_ref_b = rdm1_ref[1][ref_state - 1]
-            self.rdm1_ss = (rdm1_ref_a + rdm1_gs_a, rdm1_ref_b + rdm1_gs_b)
-        else:
-            self.rdm1_ss = rdm1_gs
-
-    def make_fno(self, rdm1_ss, thresh):
-        nocc_a = self._scf.nelec[0]
-        nocc_b = self._scf.nelec[1]
-        mo_energy_a = self.mo_energy_hf[0]
-        mo_energy_b = self.mo_energy_hf[1]
-        mo_a_coeff = self.mo_coeff_hf[0]
-        mo_b_coeff = self.mo_coeff_hf[1]
-        masks = self._mo_splitter()
-        mask_a = masks[0]
-        mask_b = masks[1]
-        rdm1_ss_a = rdm1_ss[0]
-        rdm1_ss_b = rdm1_ss[1]
-
-        n_a,V_a = np.linalg.eigh(rdm1_ss_a[nocc_a:,nocc_a:])
-        idx = np.argsort(n_a)[::-1]
-        n_a,V_a = n_a[idx], V_a[:,idx]
-        T_a = n_a > thresh
-        n_fro_vir_a = np.sum(T_a == 0)
-        T_a = np.diag(T_a)
-        V_trunc_a = V_a.dot(T_a)
-        n_keep_a = V_trunc_a.shape[0]-n_fro_vir_a
-
-        n_b,V_b = np.linalg.eigh(rdm1_ss_b[nocc_b:,nocc_b:])
-        idx = np.argsort(n_b)[::-1]
-        n_b,V_b = n_b[idx], V_b[:,idx]
-        T_b = n_b > thresh
-        n_fro_vir_b = np.sum(T_b == 0)
-        T_b = np.diag(T_b)
-        V_trunc_b = V_b.dot(T_b)
-        n_keep_b = V_trunc_b.shape[0]-n_fro_vir_b
-
-        moeoccfrz0_a, moeocc_a, moevir_a, moevirfrz0_a = [mo_energy_a[m] for m in mask_a]
-        orboccfrz0_a, orbocc_a, orbvir_a, orbvirfrz0_a = [mo_a_coeff[:,m] for m in mask_a]
-        F_can_a =  np.diag(moevir_a)
-        F_trunc_a = V_trunc_a.T.dot(F_can_a).dot(V_trunc_a)
-        e_trunc_a,Z_trunc_a = np.linalg.eigh(F_trunc_a[:n_keep_a,:n_keep_a])
-        U_vir_act_a = orbvir_a.dot(V_trunc_a[:,:n_keep_a]).dot(Z_trunc_a)
-        U_vir_fro_a = orbvir_a.dot(V_trunc_a[:,n_keep_a:])
-
-        moeoccfrz0_b, moeocc_b, moevir_b, moevirfrz0_b = [mo_energy_b[m] for m in mask_b]
-        orboccfrz0_b, orbocc_b, orbvir_b, orbvirfrz0_b = [mo_b_coeff[:,m] for m in mask_b]
-        F_can_b =  np.diag(moevir_b)
-        F_trunc_b = V_trunc_b.T.dot(F_can_b).dot(V_trunc_b)
-        e_trunc_b,Z_trunc_b = np.linalg.eigh(F_trunc_b[:n_keep_b,:n_keep_b])
-        U_vir_act_b = orbvir_b.dot(V_trunc_b[:,:n_keep_b]).dot(Z_trunc_b)
-        U_vir_fro_b = orbvir_b.dot(V_trunc_b[:,n_keep_b:])
-
-        no_comp_a = (orboccfrz0_a,orbocc_a,U_vir_act_a,U_vir_fro_a,orbvirfrz0_a)
-        no_e_comp_a = (moeoccfrz0_a, moeocc_a, e_trunc_a, moevir_a[n_keep_a:], moevirfrz0_a)
-        no_coeff_a = np.hstack(no_comp_a)
-        no_energy_a = np.hstack(no_e_comp_a)
-        nocc_loc_a = np.cumsum([0]+[x.shape[1] for x in no_comp_a]).astype(int)
-        no_frozen_a = np.hstack((np.arange(nocc_loc_a[0], nocc_loc_a[1]),
-                                np.arange(nocc_loc_a[3], nocc_loc_a[5]))).astype(int)
-
-        no_comp_b = (orboccfrz0_b,orbocc_b,U_vir_act_b,U_vir_fro_b,orbvirfrz0_b)
-        no_e_comp_b = (moeoccfrz0_b, moeocc_b, e_trunc_b, moevir_b[n_keep_b:], moevirfrz0_b)
-        no_coeff_b = np.hstack(no_comp_b)
-        no_energy_b = np.hstack(no_e_comp_b)
-        nocc_loc_b = np.cumsum([0]+[x.shape[1] for x in no_comp_b]).astype(int)
-        no_frozen_b = np.hstack((np.arange(nocc_loc_b[0], nocc_loc_b[1]),
-                                np.arange(nocc_loc_b[3], nocc_loc_b[5]))).astype(int)
-
-        no_coeff = (no_coeff_a,no_coeff_b)
-        no_energy = (no_energy_a,no_energy_b)
-        no_frozen = (no_frozen_a,no_frozen_b)
-
-        if isinstance(self._scf, scf.rohf.ROHF):
-            f_ov_a, f_ov_b = self.f_ov
-            f_ov_a = f_ov_a.dot(V_trunc_a[:,:n_keep_a]).dot(Z_trunc_a)
-            f_ov_b = f_ov_b.dot(V_trunc_b[:,:n_keep_b]).dot(Z_trunc_b)
-            f_ov = (f_ov_a,f_ov_b)
-            return no_coeff,no_energy,no_frozen,f_ov
-        else:
-            return no_coeff,no_energy,no_frozen
-
 if __name__ == '__main__':
     from pyscf import gto
     from pyscf import adc
